Fix: 메뉴로 시작안되고 바로 테트리스 시작되는 문제 해결
